requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
python-dotenv
pandas
pydantic
openpyxl
langchain-community
sentence-transformers
torch
transformers
faiss-cpu
